+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased ##
 
 * **Added:** Support for command line arguments (#18).
+* **Added:** Support for setting working directory.
 
 ## 0.3 ##